.gitignore fix
