+++ greenberry/debug_cp.py
@@ -1,0 +1,11 @@
+class Debug_cp(object):
+    '''
+    controled debugger logging module better
+    '''
+    def __init__(self, name):
+        self.name = name
+        self.var = 1
+    
+    def run(self):
+        print(self.name,'*** entered cp',self.var)
+        self.var += 1
